+++ src/main/java/com/example/coffeehouse/repositories/ClientCrudRepostirory.java
@@ -10,19 +10,19 @@ import java.math.BigDecimal;
 import java.util.List;
 
 public interface ClientCrudRepostirory extends CrudRepository<Client, Integer> {
-    public Client findById(int id);
+    public Client findById(long id);
 
     @Modifying
     @Query("update Client c set c.money = c.money + :money where c.id = :id")
-    public void updMoney(@Param("id") int id, @Param("money") BigDecimal money);
+    public void updMoney(@Param("id") long id, @Param("money") BigDecimal money);
 
     @Modifying
     @Query("update Client c set c.name = :name, c.surname = :surname, c.age = :age, c.sex = :sex where c.id = :id")
-    public void updClientGeneralInfo(@Param("id") int id, @Param("name") String name, @Param("surname") String surname, @Param("age") int age, @Param("sex") String sex);
+    public void updClientGeneralInfo(@Param("id") long id, @Param("name") String name, @Param("surname") String surname, @Param("age") int age, @Param("sex") String sex);
 
     @Modifying
     @Query("update Client c set c.email = :email, c.password = :password where c.id = :id")
-    public void updClientLogInInfo(@Param("id") int id, @Param("email") String email, @Param("password") String password);
+    public void updClientLogInInfo(@Param("id") long id, @Param("email") String email, @Param("password") String password);
 
-    public void deleteById(int id);
+    public void deleteById(long id);
 }
